refactor: extract error.rs tests to tests/error_tests.rs (#545)
Move 879 lines of inline tests from src/error.rs to a dedicated
integration test file, reducing error.rs from 1719 to 840 lines.
All test cases preserved; all pass with --no-default-features.
